separated d3 specific code from main index.js
