+++ src/index.js
@@ -5,11 +5,14 @@ import "../js/leaflet.heat";
 import "../js/leaflet-slider";
 import select2 from "../js/select2.min.js";
 import "../js/L.Control.Range";
+import { updateGraph, initPlotChart } from "./plot";
+import { strongHeader, strongLine } from "./utility";
 
 // Globals ---------------------------------------------------------------------------
 window.circles = [];
 window.contextTags = [];
 window.siteAggregates = {};
+window.otuLookup = {};
 window.sampleContextLookup = {};
 
 const dev_url = "http://localhost:8000/edna/api/v1.0/";
@@ -29,34 +32,10 @@ const API_URLS = {
   otu_code_by_id: dev_url + "otu/?id="
 };
 
-// helper functions ------------------------------------------------------------------
-
-function round(x, dp) {
-  var factor = Math.pow(10, dp);
-  var tmp = x * factor;
-  tmp = Math.round(tmp);
-  return tmp / factor;
-}
-
-/**
- * Simple helper to wrap contents in strong tags followed by a line break.
- */
-const strongLine = s => {
-  return "<strong>" + s + "</strong>" + "<br />";
-};
-
-/**
- * Helper function to make a strong field label with a value.
- */
-const strongHeader = (header, text) => {
-  return "<strong>" + header + ": </strong> " + text + "<br />";
-};
-
 /**
  * Gets all filter element values and uses them as parameters for an API request.
  */
 function fetchSampleOtus() {
-  console.log("fet sample otu call");
   let contextParams = $("#select-contextual").select2("data");
   let taxonParams = $("#select-taxonomic").select2("data");
   // Formatting and adding otu arguments
@@ -384,7 +363,7 @@ function makePopupContent(properties, jsonResponse = null) {
   //list all sites within the cell.
   popupContent += strongLine("Sites in cell: ") + "<ul>";
   for (let i in properties.sites) {
-    siteId = properties.sites[i];
+    let siteId = properties.sites[i];
     popupContent += "<li>" + window.sampleContextLookup[siteId].site + "</li>";
   }
   popupContent += "</ul><br />";
@@ -476,7 +455,7 @@ function featureCollectionToLayer(featureCollection, property, layerGroup) {
       if (missingIds.length > 0) {
         // f_url = "http://localhost:8000/edna/api/v1.0/otu/?";
         // f_url += "id=" + missingIds.join("&id=");
-        f_url = API_URLS.otu_code_by_id + missingIds.join("&id=");
+        let f_url = API_URLS.otu_code_by_id + missingIds.join("&id=");
         console.log(f_url);
         fetch(f_url).then(response => {
           response.json().then(jsonResponse => {
@@ -815,7 +794,7 @@ function highlightFeatureClick(layer) {
  * the need for mouseover/mouse click on the layer
  * @param {*} layer
  */
-function highlightLayer(layer) {
+export function highlightLayer(layer) {
   layer.setStyle({
     weight: 5,
     opacity: 0.9
@@ -830,7 +809,7 @@ function highlightLayer(layer) {
  * Values are hardcoded due to geojson.reset() not working as planned.
  * @param {*} layer
  */
-function disableHighlightLayer(layer) {
+export function disableHighlightLayer(layer) {
   var properties = layer.feature.properties;
   //console.log(properties);
   layer.setStyle({
@@ -856,230 +835,6 @@ function addSiteMetrics(siteAggregates) {
     site.effectiveAlpha = Math.exp(site.shannonDiversity);
   }
   return siteAggregates;
-}
-
-/**
- * Calculates queries max and minimum site metrics.
- * Returns colour range with spectrum from minimum value metric to max value metric.
- * @param {*} metric
- * @param {*} siteAggregates
- */
-function createColorRange(siteAggregates) {
-  // gets value from drop down and creates colour scale from the select option.
-  const metric = document.getElementById("meta-select").value;
-  const sites = [];
-  for (var siteId in siteAggregates) {
-    sites.push(window.sampleContextLookup[siteId]);
-  }
-  const min = d3.min(sites, function(d) {
-    return d[metric];
-  });
-  const max = d3.max(sites, function(d) {
-    return d[metric];
-  });
-  console.log("visualization plot min, max:", min, max);
-  const colorScheme = document.getElementById("colour-scheme-select").value;
-  let colorRange = [];
-  switch (colorScheme) {
-    case "sequential":
-      colorRange = ["blue", "orange"];
-      break;
-    case "diverging":
-      colorRange = ["#2c7bb6", "#d7191c"];
-      break;
-    default:
-      colorRange = ["grey", "black"];
-  }
-  const colourRange = d3
-    .scaleLinear()
-    .interpolate(d3.interpolateRgb)
-    .domain([min, max])
-    .range(colorRange);
-  return colourRange;
-}
-
-/**
- * Returns a random amount between upper and lower. For jittering the plots.
- */
-let randomRange = (upper, lower) => {
-  return Math.random() * (upper - lower) + lower;
-};
-
-/**
- * Converts siteAggregates to an easier format for d3 use. Updates existing datapoints, enters new additional datapoints
- * @param {*} siteAggregates
- */
-function updateGraph(siteAggregates) {
-  // todo: see if I can make this into one class. Called in colorrange, select onchange function as well.
-  var metricColour = createColorRange(siteAggregates);
-  var colourMetric = document.getElementById("meta-select").value;
-
-  function makeNestableObject(siteId, site, metricName, value) {
-    return {
-      siteId: siteId,
-      Metric: metricName,
-      value: value,
-      meta: sampleContextLookup[siteId]
-    };
-  }
-
-  var plotData = [];
-  for (const [siteId, site] of Object.entries(siteAggregates)) {
-    plotData.push(
-      makeNestableObject(siteId, site, "OTU richness", site.richness)
-    );
-    plotData.push(
-      makeNestableObject(siteId, site, "Shannon entropy", site.shannonDiversity)
-    );
-    plotData.push(
-      makeNestableObject(siteId, site, "Sequence abundance", site.abundance)
-    );
-    plotData.push(
-      makeNestableObject(
-        siteId,
-        site,
-        "Effective alpha diversity",
-        site.effectiveAlpha
-      )
-    );
-  }
-
-  var nestedPlotData = d3
-    .nest()
-    .key(function(d) {
-      return d.Metric;
-    })
-    .entries(plotData);
-
-  //within the svg, within the g tags, select the class datapoints
-  var update = g.selectAll(".datapoints").data(nestedPlotData, function(d) {
-    return d.values;
-  });
-  update
-    .enter()
-    .append("g")
-    .attr("class", "datapoints")
-    .merge(update)
-    .each(function(d) {
-      //loop through each data group
-      const min = d3.min(d.values, function(d) {
-        return d.value;
-      });
-      const max = d3.max(d.values, function(d) {
-        return d.value;
-      });
-      const mean = d3.mean(d.values, function(d) {
-        return d.value;
-      });
-      //console.log(min, max, mean);
-      var circle = d3
-        .select(this)
-        .selectAll("circle")
-        .data(d.values, function(d) {
-          return d.siteId;
-        });
-      circle.exit().remove();
-      //Enter statement
-      circle
-        .enter()
-        .append("circle")
-        .attr("class", "enter")
-        .attr("id", d => d.meta.site)
-        //.attr('cy', y(d.key))
-        .attr("cy", function(circle) {
-          // * If no jitter wanted then set jitter 0, 0.
-          return y(circle.Metric) + randomRange(10, -10);
-        })
-        .attr("r", 7)
-        .attr("opacity", 0.3)
-        .attr("fill", function(d) {
-          //TODO: Create a function to get the select dropdown values. For here and onchange.
-          //console.log(d.meta[metric]);
-          return metricColour(d.meta[colourMetric]);
-        })
-        .on("mouseover", function(d) {
-          d3.select(this.parentNode.parentNode)
-            .selectAll("#" + d.meta.site)
-            .transition()
-            .attr("r", 14)
-            .duration(250);
-          tooltip
-            .transition()
-            .style("opacity", 0.9)
-            .duration(250);
-          tooltip
-            .html(
-              strongLine(d.meta.site) +
-                strongHeader(d.Metric, d.value) +
-                strongHeader(
-                  document.getElementById("meta-select").value,
-                  // d.meta[document.getElementById("meta-select").value]
-                  d.meta["elev"]
-                )
-            )
-            .style("left", d3.event.pageX + "px")
-            .style("top", d3.event.pageY - 10 + "px")
-            .style("opacity", 0.9)
-            .style("z-index", 1000);
-          let layer = GetLayerBySampleContextId(d.siteId);
-          if (layer != null && layer !== undefined) {
-            highlightLayer(layer);
-          }
-        })
-        .on("mouseout", function(d) {
-          d3.select(this.parentNode.parentNode)
-            .selectAll("#" + d.meta.site)
-            .transition()
-            .attr("r", 7)
-            .duration(250);
-          tooltip
-            .transition()
-            .style("opacity", 0)
-            .style("z-index", 1000)
-            .duration(250);
-          let layer = GetLayerBySampleContextId(d.siteId);
-          if (layer != null && layer !== undefined) {
-            disableHighlightLayer(layer);
-          }
-        })
-        .on("click", function(d) {
-          let layer = GetLayerBySampleContextId(d.siteId);
-          if (layer != null && layer !== undefined) {
-            let bounds = layer.getBounds();
-            map.flyToBounds(bounds, { padding: [300, 300] });
-          }
-        })
-        .merge(circle)
-        .transition()
-        .duration(1500)
-        .attr("cx", function(d) {
-          var cx;
-          if (max == min) {
-            cx = 0;
-          } else {
-            cx = (d.value - min) / (max - min);
-          }
-          return x(cx);
-        });
-
-      //adding mean circles
-      d3.select(this)
-        .append("circle")
-        .attr("class", "enter-mean")
-        .attr("cy", y(d.key))
-        .attr("r", 15)
-        .style("stroke", "grey")
-        .style("stroke-width", 2)
-        .style("fill", "none")
-        .style("opacity", 0)
-        .transition()
-        .duration(1500)
-        .style("opacity", 0.75)
-        .attr("cx", x((mean - min) / (max - min)));
-    }); //.each() end.
-
-  // should remove extras. Not sure if necessary.
-  update.exit().remove();
 }
 
 //generating the map
@@ -1119,7 +874,6 @@ var mode = taxonParams.get("mode");
 
 var detailLevel = 60;
 //warrick map additions
-var grid = makeGrid(detailLevel);
 
 //shows the scale of the map
 var scaleIndicator = L.control.scale().addTo(map);
@@ -1151,7 +905,7 @@ var layerMenuControl = L.control
  * Takes in a sample contextual index, iterates through all active and stamped map layers, then returns the last matching layer with leaflet id. Assumes sample contextual contains a leafletId value.
  * @param {integer} id
  */
-function GetLayerBySampleContextId(id) {
+export function GetLayerBySampleContextId(id) {
   let targetLeafletIds = window.sampleContextLookup[id].leafletIds;
   // for now just iterate through all layers. Final layer that matches is returned.
   let targetLayer;
@@ -1186,7 +940,7 @@ function changeSliderValue(value) {
  */
 
 function initializeDisplayOptionsControl() {
-  let slider = L.control.slider(
+  let gridResSlider = L.control.slider(
     function(value) {
       detailLevel = value;
       // TODO: directly call fetchSampleOTU data instead of triggering a change.
@@ -1207,7 +961,18 @@ function initializeDisplayOptionsControl() {
       syncSlider: true
     }
   );
-  slider.addTo(map);
+  gridResSlider.addTo(map);
+
+  let gridResInput = L.control({
+    position: "bottomleft"
+  });
+  gridResInput.onAdd = map => {
+    let _div = L.DomUtil.create("div", "info");
+    _div.innerHTML =
+      '<label for="grid-input">Grid Resolution: </label><input id="grid-input" placeholder="Type value" type="number" onchange="changeSliderValue(this.value)"/>';
+    return _div;
+  };
+  gridResInput.addTo(map);
 
   let displayControlRoot = L.control({
     position: "bottomleft"
@@ -1222,9 +987,12 @@ function initializeDisplayOptionsControl() {
     togglableElements.className = "display-controls-togglable";
     displayControlRoot.appendChild(togglableElements);
 
-    togglableElements.appendChild(slider.getContainer());
-    togglableElements.appendChild(gridResolutionInput.getContainer());
+    // adding pre-existing elements that will be hidden by toggling.
+    togglableElements.appendChild(gridResSlider.getContainer());
+    togglableElements.appendChild(gridResInput.getContainer());
     togglableElements.appendChild(layerMenuControl.getContainer());
+
+    // adding the toggle button
     let displayControlVisibleButton = L.DomUtil.create(
       "button",
       "display-controls-root-button"
@@ -1243,18 +1011,6 @@ function initializeDisplayControlButton() {
     $("#display-controls-togglable").toggle("slow");
   };
 }
-
-let gridResolutionInput = L.control({
-  position: "bottomleft"
-});
-gridResolutionInput.onAdd = map => {
-  let _div = L.DomUtil.create("div", "info");
-  // todo: Shrink down the input field to 4/5 numbers.
-  _div.innerHTML =
-    '<label for="grid-input">Grid Resolution: </label><input id="grid-input" placeholder="Type value" type="number" onchange="changeSliderValue(this.value)"/>';
-  return _div;
-};
-gridResolutionInput.addTo(map);
 
 //Adding custom control for Andrew's Visualization Copy.
 const leafletGraphControl = L.control({ position: "bottomright" });
@@ -1299,75 +1055,11 @@ function updateGraphColours(metric) {
 }
 
 //Adding d3 visualization
-const { g, y, tooltip, x } = createGraph();
+export const { g, y, tooltip, x } = initPlotChart();
 
 var hashComponents = decodeURIComponent(
   window.location.hash.replace("#", "")
 ).split(",");
-
-function createGraph() {
-  var margin = { top: 20, right: 30, bottom: 20, left: 160 },
-    width = window.innerWidth * 0.75 - margin.left - margin.right,
-    height = window.innerHeight * 0.35 - margin.top - margin.bottom;
-  var chart = d3
-    .select("#chart")
-    .append("svg")
-    .attr("width", width + margin.right + margin.left)
-    .attr("height", height + margin.top + margin.bottom);
-  var main = chart
-    .append("g")
-    .attr("transform", "translate(" + margin.left + "," + margin.top + ")")
-    .attr("width", width)
-    .attr("height", height)
-    .attr("id", "main");
-  var x = d3
-    .scaleLinear()
-    .domain([0, 1]) // the range of the values to plot
-    .range([0, width]); // the pixel range of the x-axis
-  var xTicks = [0, 1];
-  var xLabels = ["Minimum", "Maximum"];
-  var xAxis = d3
-    .axisBottom()
-    .scale(x)
-    .tickValues([0, 1])
-    .tickFormat(function(d, i) {
-      return xLabels[i];
-    });
-  var y = d3
-    .scalePoint()
-    //.domain(nested.map( function(d) { return d.key }) )
-    .domain([
-      "OTU richness",
-      "Sequence abundance",
-      "Shannon entropy",
-      "Effective alpha diversity",
-      "Orders"
-    ])
-    .range([0, height - 20])
-    .padding(0.1);
-  var yAxis = d3.axisLeft().scale(y);
-  // Draw the x and y axes
-  main
-    .append("g")
-    .attr("transform", "translate(0," + height + ")") // Position at bottom of chart
-    .attr("class", "main axis")
-    .attr("id", "xAxis")
-    .call(xAxis);
-  main
-    .append("g")
-    .attr("transform", "translate(0,0)") // Position at left of chart
-    .attr("class", "main axis")
-    .attr("id", "yAxis")
-    .call(yAxis);
-  // Draw the graph object holding the data points
-  var g = main.append("svg:g").attr("id", "datapoints");
-  var tooltip = d3
-    .select("#map")
-    .append("div")
-    .attr("class", "tooltip")
-    .style("opacity", 0);
-  return { g, y, tooltip, x };
-}
 
 function createLoadingMessage() {
   let popupDiv = document.createElement("div");
