+++ testsToMake.md
@@ -1,0 +1,11 @@
+test changing select then getting heat results.
+
+test clicking on a grid cell to load popup content.
+
+test hovering on plot point generates tooltip.
+
+test hover on plot generates grid outline.
+
+test click on plot point flies to grid cell.
+
+test changing grid res updates gridcells
